feat(express-server): handle listen startup errors and improve graceful shutdown
> this PR started out as handling the err arg to `app.listen` but I threw in some other graceful shutdown niceities, optional but here for your consideration

In Express v5, `app.listen`'s callback can accept an err arg, in case
the server emitted an error during startup.

You can handle that and fall into the graceful shutdown in that (rare)
case.

For the graceful shutdown, I promisified the `server.close` to improve
control flow there, and added in an exit code arg. Also used the Node
18+ feature of [`server.closeIdleConnections()`](https://nodejs.org/api/http.html#servercloseidleconnections)
to drain any connections idling for a quick exit.
That is pure back compat, as starting in node 19+, `server.close` will also effectively
`closeIdleConnections`. This ensures the lower end of
your support range (Node 18) behaves the same as 19 and up.

On the SIG listeners, its simpler to just throw a `once` on them than to
track if we are already in a shutdown scenario, in the case that
mutliple signals race.

The timer `unref` is more defense in depth here, as long as you cleanup
the timer it wont hold the event loop open. But for this specific kind
of timer, no reason for it to keep going if the event loop is otherwise
empty

I set `HOST` as a var just as a convenience for the logging.